Moved scripts, minor fixes, added flags
- Moved the UI scripts to the UI folder (oops)
- Added flags (both UI and physical, as placeholders)
