+++ Assets/Scripts/UI Scripts/Game UI Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b1e4522ab84e24927b0202bfe5443a23
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
